+++ R/RcppExports.R
@@ -208,6 +208,10 @@ rewire_swap <- function(graph, nsteps = 100L, self = FALSE, multiple = FALSE, un
     .Call('netdiffuseR_rewire_swap', PACKAGE = 'netdiffuseR', graph, nsteps, self, multiple, undirected, pr_rewire, althexagons)
 }
 
+rgraph_er_cpp <- function(n = 10L, p = 0.3, undirected = TRUE, weighted = FALSE, self = FALSE) {
+    .Call('netdiffuseR_rgraph_er_cpp', PACKAGE = 'netdiffuseR', n, p, undirected, weighted, self)
+}
+
 rgraph_ba_cpp <- function(graph, dgr, m = 1L, t = 10L, self = TRUE) {
     .Call('netdiffuseR_rgraph_ba_cpp', PACKAGE = 'netdiffuseR', graph, dgr, m, t, self)
 }
@@ -216,8 +220,12 @@ rgraph_ba_new_cpp <- function(m0 = 1L, m = 1L, t = 10L, self = TRUE) {
     .Call('netdiffuseR_rgraph_ba_new_cpp', PACKAGE = 'netdiffuseR', m0, m, t, self)
 }
 
-rgraph_er_cpp <- function(n = 10L, p = 0.3, undirected = TRUE, weighted = FALSE, self = FALSE) {
-    .Call('netdiffuseR_rgraph_er_cpp', PACKAGE = 'netdiffuseR', n, p, undirected, weighted, self)
+rgraph_sf_homo_cpp <- function(eta, graph, dgr, m = 1L, t = 10L, self = TRUE) {
+    .Call('netdiffuseR_rgraph_sf_homo_cpp', PACKAGE = 'netdiffuseR', eta, graph, dgr, m, t, self)
+}
+
+rgraph_sf_homo_new_cpp <- function(eta, m0 = 1L, m = 1L, t = 10L, self = TRUE) {
+    .Call('netdiffuseR_rgraph_sf_homo_new_cpp', PACKAGE = 'netdiffuseR', eta, m0, m, t, self)
 }
 
 #' Ring lattice graph